requirements.txt
fastapi==0.114.1
uvicorn[standard]==0.30.6
pandas>=1.0.3,<2.0.0
python-dotenv==1.0.1
opensoundscape==0.6.0
python-telegram-bot==21.5
joblib==1.3.2
requests==2.31.0
torch==1.13.1
torchvision==0.14.1
python-multipart==0.0.6